+++ crates/yagra-core/src/api/mod.rs
@@ -858,6 +858,7 @@ mod tests {
             sys_descr: None,
             os_version: None,
             os_version_without_patch: None,
+            serial_number: None,
             sys_object_id: None,
             dns_chain: None,
             neighbors: None,