+++ crates/yagra-poller/src/worker/interfaces.rs
@@ -264,6 +264,7 @@ async fn execute_table_walk(
         sys_descr: None,
         os_version: None,
         os_version_without_patch: None,
+        serial_number: None,
         sys_object_id: None,
         dns_chain: None,
         neighbors: None,