+++ crates/yagra-poller/src/worker/meraki.rs
@@ -90,6 +90,7 @@ pub async fn execute_meraki(
             sys_descr: None,
             os_version: None,
             os_version_without_patch: None,
+            serial_number: None,
             sys_object_id: None,
             dns_chain: None,
             neighbors: None,